updated Theme.kt
