+++ core/designsystem/src/commonMain/kotlin/org/mifospay/core/designsystem/theme/Type.kt
@@ -9,6 +9,7 @@
  */
 package org.mifospay.core.designsystem.theme
 
+import androidx.compose.material3.Typography
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
@@ -26,10 +27,9 @@ import mobile_wallet.core.designsystem.generated.resources.outfit_regular
 import mobile_wallet.core.designsystem.generated.resources.outfit_semi_bold
 import mobile_wallet.core.designsystem.generated.resources.outfit_thin
 import org.jetbrains.compose.resources.Font
-import template.core.base.designsystem.theme.KptTypographyImpl
 
 @Composable
-internal fun getFontFamily(): FontFamily {
+private fun fontFamily(): FontFamily {
     return FontFamily(
         Font(Res.font.outfit_black, FontWeight.Black),
         Font(Res.font.outfit_bold, FontWeight.Bold),
@@ -43,38 +43,46 @@ internal fun getFontFamily(): FontFamily {
     )
 }
 
-internal val mifosTypography = KptTypographyImpl(
+// Set of Material typography styles to start with
+@Composable
+internal fun getTypography() = Typography(
     displayLarge = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 57.sp,
         lineHeight = 64.sp,
         letterSpacing = (-0.25).sp,
     ),
     displayMedium = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 45.sp,
         lineHeight = 52.sp,
         letterSpacing = 0.sp,
     ),
     displaySmall = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 36.sp,
         lineHeight = 44.sp,
         letterSpacing = 0.sp,
     ),
     headlineLarge = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 32.sp,
         lineHeight = 40.sp,
         letterSpacing = 0.sp,
     ),
     headlineMedium = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 28.sp,
         lineHeight = 36.sp,
         letterSpacing = 0.sp,
     ),
     headlineSmall = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 24.sp,
         lineHeight = 32.sp,
@@ -85,17 +93,20 @@ internal val mifosTypography = KptTypographyImpl(
         ),
     ),
     titleLarge = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.SemiBold,
         fontSize = 24.sp,
         lineHeight = 30.24.sp,
     ),
     titleMedium = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.SemiBold,
         fontSize = 20.sp,
         lineHeight = 28.sp,
         letterSpacing = 0.1.sp,
     ),
     titleSmall = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Medium,
         fontSize = 14.sp,
         lineHeight = 20.sp,
@@ -103,6 +114,7 @@ internal val mifosTypography = KptTypographyImpl(
     ),
     // Default text style
     bodyLarge = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 16.sp,
         lineHeight = 24.sp,
@@ -113,12 +125,14 @@ internal val mifosTypography = KptTypographyImpl(
         ),
     ),
     bodyMedium = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 14.sp,
         lineHeight = 20.sp,
         letterSpacing = 0.25.sp,
     ),
     bodySmall = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Normal,
         fontSize = 12.sp,
         lineHeight = 16.sp,
@@ -126,6 +140,7 @@ internal val mifosTypography = KptTypographyImpl(
     ),
     // Used for Button
     labelLarge = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Medium,
         fontSize = 16.sp,
         lineHeight = 20.sp,
@@ -133,6 +148,7 @@ internal val mifosTypography = KptTypographyImpl(
     ),
     // Used for Navigation items
     labelMedium = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Medium,
         fontSize = 12.sp,
         lineHeight = 16.sp,
@@ -144,6 +160,7 @@ internal val mifosTypography = KptTypographyImpl(
     ),
     // Used for Tag
     labelSmall = TextStyle(
+        fontFamily = fontFamily(),
         fontWeight = FontWeight.Medium,
         fontSize = 10.sp,
         lineHeight = 14.sp,
